requirements.txt
absl-py==0.7.1
astor==0.8.0
Click==7.0
cycler==0.10.0
decorator==4.4.0
Flask==1.1.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
gunicorn==19.9.0
h5py==2.9.0
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.3
numpy==1.16.4
opencv-python==3.3.0.9
Pillow==6.1.0
protobuf==3.9.0
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1.1
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.15.5
wrapt==1.11.2
